+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-## [0.6.0] - 2017-03-21
+## [0.6.0] - 2017-03-23
 
 ### Added
 
 - 3PlayMedia support. Now you can fetch transcripts for your video from 3PM.
+- Open edX settings support. Now administrators can set default values
+  for VideoXBlock on a system-wide level.
 
 ### Changed